// src/app/modules/planetas/models/planeta.ts
export class Planeta {
  id: number;
  nombre: string;
  periodo_rotacion: number;
  diametro: number;
  clima: number;
  terreno: string;
  masa: number;
  radio_orbital: number;
  personas: any;
  contador: number;
  deleted_at: string;
  created_at: string;
  updated_at: string;

}


JSON.stringify